Fix crashes found during fuzz testing (#70)
fixes issue #69 - crash on "DATA\r\n"

This one fixes a crash when the first input is "DATA\r\n"
The fix was to change envelope.EmailAddress to be a value rather than a pointer.
The reason is that the struct is small and consists of slices anyway, so not much is saved when passing by pointer. Also, it doesn't not mutate after being passed.